+++ back-end/modules/Services/log/user/src/services/userService.ts
@@ -1,4 +1,4 @@
-import { User } from './../models/user';
+import { User } from '../models/user';
 import bcrypt from 'bcrypt';
 import axios from 'axios';
