+++ back-end/modules/Services/work/artwork/src/services/artworkService.ts
